# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FortKnoxx Security Intelligence Platform
  description: 'Unified security scanning API: SAST, DAST, SCA, IaC, secrets, container, and AI-powered scanners (zero-day
    GNN, business logic, LLM adversarial, auth).'
  version: 1.0.0
paths:
...
  /api/repositories:
    get:
      tags:
      - repositories
      summary: List Repositories
      operationId: list_repositories_api_repositories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Repository'
                type: array
                title: Response List Repositories Api Repositories Get
    post:
      summary: Add Repository
      description: Add a repository for scanning (supports both public repos and private repos with tokens)
      operationId: add_repository_api_repositories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRepositoryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/vulnerabilities/{scan_id}:
    get:
      tags:
      - findings
      summary: List Vulnerabilities For Scan
      operationId: list_vulnerabilities_for_scan_api_vulnerabilities__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vulnerability'
                title: Response List Vulnerabilities For Scan Api Vulnerabilities  Scan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/vulnerabilities/repo/{repo_id}:
    get:
      tags:
      - findings
      summary: List Vulnerabilities For Repo
      operationId: list_vulnerabilities_for_repo_api_vulnerabilities_repo__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vulnerability'
                title: Response List Vulnerabilities For Repo Api Vulnerabilities Repo  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/quality/{scan_id}:
    get:
      tags:
      - findings
      summary: List Quality Issues For Scan
      operationId: list_quality_issues_for_scan_api_quality__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Quality Issues For Scan Api Quality  Scan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/quality/repo/{repo_id}:
    get:
      tags:
      - findings
      summary: List Quality Issues For Repo
      operationId: list_quality_issues_for_repo_api_quality_repo__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Quality Issues For Repo Api Quality Repo  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/quality/summary/{scan_id}:
    get:
      tags:
      - findings
      summary: Quality Summary
      operationId: quality_summary_api_quality_summary__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Quality Summary Api Quality Summary  Scan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/compliance/{scan_id}:
    get:
      tags:
      - findings
      summary: List Compliance Issues For Scan
      operationId: list_compliance_issues_for_scan_api_compliance__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Compliance Issues For Scan Api Compliance  Scan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/compliance/repo/{repo_id}:
    get:
      tags:
      - findings
      summary: List Compliance Issues For Repo
      operationId: list_compliance_issues_for_repo_api_compliance_repo__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Compliance Issues For Repo Api Compliance Repo  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/compliance/summary/{scan_id}:
    get:
      tags:
      - findings
      summary: Compliance Summary
      operationId: compliance_summary_api_compliance_summary__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Compliance Summary Api Compliance Summary  Scan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sbom/{repo_id}:
    get:
      tags:
      - findings
      summary: Get Sbom
      operationId: get_sbom_api_sbom__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Sbom Api Sbom  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/integrations/git/connect:
    post:
      tags:
      - integrations
      summary: Connect Integration
      operationId: connect_integration_api_integrations_git_connect_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectGitIntegrationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Connect Integration Api Integrations Git Connect Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/integrations/git/{provider}:
    delete:
      tags:
      - integrations
      summary: Disconnect Integration
      operationId: disconnect_integration_api_integrations_git__provider__delete
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Disconnect Integration Api Integrations Git  Provider  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/integrations/git/{provider}/repositories:
    get:
      tags:
      - integrations
      summary: List Remote Repositories
      operationId: list_remote_repositories_api_integrations_git__provider__repositories_get
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 30
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Remote Repositories Api Integrations Git  Provider  Repositories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AddRepositoryRequest:
      properties:
        provider:
          $ref: '#/components/schemas/GitProvider'
        repo_url:
          type: string
          title: Repo Url
          description: Repository URL (e.g., https://github.com/owner/repo)
        auto_scan:
          type: boolean
          title: Auto Scan
          description: Enable auto-scan on push
          default: false
        branch:
          anyOf:
          - type: string
          - type: 'null'
          title: Branch
          description: Branch to scan (defaults to default branch)
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
          description: Access token for private repos (optional for public repos)
        is_public:
          type: boolean
          title: Is Public
          description: Whether this is a public repository (no token required)
          default: false
      type: object
      required:
      - provider
      - repo_url
      title: AddRepositoryRequest
      description: Request model for adding a repository to scan
      example:
        auto_scan: true
        branch: main
        is_public: true
        provider: github
        repo_url: https://github.com/myorg/myrepo
...
    ConnectGitIntegrationRequest:
      properties:
        provider:
          $ref: '#/components/schemas/GitProvider'
        name:
          type: string
          title: Name
          description: Display name for this integration
        access_token:
          type: string
          title: Access Token
          description: Personal access token
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
          description: Base URL for self-hosted (e.g., https://gitlab.company.com)
      type: object
      required:
      - provider
      - name
      - access_token
      title: ConnectGitIntegrationRequest
      description: Request model for connecting a Git provider
      example:
        access_token: ghp_xxxxxxxxxxxx
        name: My GitHub Account
        provider: github
...
    GitProvider:
      type: string
      enum:
      - github
      - gitlab
      title: GitProvider
      description: Supported Git providers
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Repository:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
        branch:
          type: string
          title: Branch
          default: main
        last_scan:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Scan
        scan_status:
          type: string
          title: Scan Status
          default: pending
        created_at:
          type: string
          format: date-time
          title: Created At
        security_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Security Score
        vulnerabilities_count:
          type: integer
          title: Vulnerabilities Count
          default: 0
        critical_count:
          type: integer
          title: Critical Count
          default: 0
        high_count:
          type: integer
          title: High Count
          default: 0
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
      type: object
      required:
      - name
      - url
      title: Repository
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Vulnerability:
      properties:
        id:
          type: string
          title: Id
        repo_id:
          type: string
          title: Repo Id
        scan_id:
          type: string
          title: Scan Id
        file_path:
          type: string
          title: File Path
        line_start:
          type: integer
          title: Line Start
        line_end:
          type: integer
          title: Line End
        severity:
          type: string
          title: Severity
        category:
          type: string
          title: Category
        owasp_category:
          type: string
          title: Owasp Category
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        code_snippet:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Snippet
          default: ''
        cwe:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe
        cvss_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Cvss Score
        fix_recommendation:
          anyOf:
          - type: string
          - type: 'null'
          title: Fix Recommendation
        detected_by:
          type: string
          title: Detected By
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - repo_id
      - scan_id
      - file_path
      - line_start
      - line_end
      - severity
      - category
      - owasp_category
      - title
      - description
      - detected_by
      title: Vulnerability